Add partial support for PREP_DIR as cmdline arg
The option `--prep-dir` is added to OnlineOptions with a default value
of "Player-Data" as with the macro. When using the command-line option,
setting the macro via the CFLAGS in the CONFIG file is not necessary.

Some parts of the codebase still depend on the PREP_DIR macro, such as
FHE*, ECDSA, and some components under Utils/.